+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [12.1.1] - 2024-02-22
+
+### Fixed
+- Handle default system font weights for SwiftUI
+- Add largeTitle style in font tests
+
 ## [12.1.0] - 2024-01-15
 
 ### Added